runpod_app/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
torch==1.12.0
torchvision==0.13.0
transformers==4.21.0
accelerate==0.20.3
bitsandbytes==0.39.0
Pillow==9.2.0
python-multipart==0.0.6
numpy==1.21.5
requests==2.28.1
